Fix NPE in MeetTogether
09.02.2022

- fix: correct MeetTogether's constructor
- change: correct codesmells

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/task/meta/PrepareDessertMeta.java
@@ -32,7 +32,7 @@ public class PrepareDessertMeta extends MetaTask {
             "Task.description.prepareDessertMeta"); //$NON-NLS-1$
 
     private static final int PREP_TIME = 10;
-    private static final int VALUE = 10;
+    private static final double VALUE = 10;
     private static final int MOD = 5;
     
     public PrepareDessertMeta() {
